Enabling auto-reformatting on commit.
We are using `friendsofphp/php-cs-fixer` and `ccinn/husky-php` in order
to run auto-formatting on every commit.

Beware that we are stuck at `php-cs-fixer@2.X.X` because `husky-php` has
some outdated dependency. We might want to migrate from this to another
solution but I didn't want to use any `npm` or `yarn` based.